chore

+++ src/lib/Metatool.WindowsInput/MouseSimulator.cs
@@ -162,16 +162,16 @@ namespace Metatool.WindowsInput
             var optimalPixelsPerMillisecond = 1;
             var minDuration = 200;
             var maxDuration = 500;
-            var duration = Convert.ToInt32(totalDistance / optimalPixelsPerMillisecond).Clamp(minDuration, maxDuration);
+            var duration = Convert.ToInt64(totalDistance / optimalPixelsPerMillisecond).Clamp(minDuration, maxDuration);
 
             // Calculate the steps for the smoothness
             var optimalPixelsPerStep = 10;
             var minSteps = 10;
             var maxSteps = 50;
-            var steps = Convert.ToInt32(totalDistance / optimalPixelsPerStep).Clamp(minSteps, maxSteps);
+            var steps = Convert.ToInt64(totalDistance / optimalPixelsPerStep).Clamp(minSteps, maxSteps);
 
             // Calculate the interval and the step size
-            var interval = duration / steps;
+            var interval = (int)(duration / steps);
             var stepX = (double)(newX - startX) / steps;
             var stepY = (double)(newY - startY) / steps;
